recommendation-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
numpy==2.2.6
scikit-learn==1.7.0
redis==5.2.1
